restructuring of project
