+++ package.json
@@ -16,29 +16,29 @@
 		"preinstall": "npx only-allow pnpm"
 	},
 	"devDependencies": {
-		"@playwright/test": "^1.41.2",
-		"@skeletonlabs/skeleton": "^2.8.0",
+		"@playwright/test": "^1.42.1",
+		"@skeletonlabs/skeleton": "^2.9.0",
 		"@skeletonlabs/tw-plugin": "^0.2.4",
 		"@sveltejs/adapter-auto": "^2.1.1",
 		"@sveltejs/adapter-node": "^1.3.1",
 		"@sveltejs/kit": "^1.30.4",
 		"@tailwindcss/forms": "^0.5.7",
-		"@types/node": "^20.11.20",
+		"@types/node": "^20.11.26",
 		"@typescript-eslint/eslint-plugin": "^6.21.0",
 		"@typescript-eslint/parser": "^6.21.0",
-		"autoprefixer": "^10.4.17",
+		"autoprefixer": "^10.4.18",
 		"eslint": "^8.57.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.35.1",
 		"postcss": "^8.4.35",
 		"postcss-load-config": "^4.0.2",
 		"prettier": "^3.2.5",
-		"prettier-plugin-svelte": "^3.2.1",
+		"prettier-plugin-svelte": "^3.2.2",
 		"svelte": "^4.2.12",
-		"svelte-check": "^3.6.4",
+		"svelte-check": "^3.6.6",
 		"tailwindcss": "^3.4.1",
 		"tslib": "^2.6.2",
-		"typescript": "^5.3.3",
+		"typescript": "^5.4.2",
 		"vite": "^4.5.2",
 		"vitest": "^0.32.4"
 	},
@@ -46,6 +46,7 @@
 	"dependencies": {
 		"@floating-ui/dom": "^1.6.3",
 		"@fortawesome/fontawesome-free": "^6.5.1",
+		"@supabase/supabase-js": "^2.39.8",
 		"pocketbase": "^0.19.0"
 	}
 }
