+++ src/lib/components/FichaSm.svelte
