+++ src/lib/utils/supabaseClient.ts
@@ -1,0 +1,3 @@
+import { createClient } from '@supabase/supabase-js';
+
+export const supabase = createClient('https://<project>.supabase.co', '<your-anon-key>');
