+++ src/routes/+page.server.ts
@@ -1,0 +1,10 @@
+// TODO: Query data from Supabase
+/* import { supabase } from '$lib/utils/supabaseClient';
+
+export async function load() {
+	const { data } = await supabase.from('countries').select();
+	return {
+		countries: data ?? []
+	};
+}
+*/
