+++ src/routes/home/+page.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
 	import { error, redirect } from '@sveltejs/kit';
 	import { goto } from '$app/navigation';
-	import { pb } from '$lib/pocketbase';
+	import { pb } from '$lib/utils/pocketbase';
 	import { browser } from '$app/environment';
-	import { currentUser } from '$lib/pocketbase';
-	import Record from '$lib/Record.svelte';
+	import { currentUser } from '$lib/utils/pocketbase';
+	import Record from '$lib/components/Record.svelte';
 	import { onMount } from 'svelte';
 
 	/*onMount(() => {
